# Optimized Min, Max and  SequenceEqual.