Min, Max, SequenceEqualを最適化しました。
Intrinsicsは消しつつ、AVX2環境で
int Min Avg 0.76 μs ->0.24 μs
Int  SequenceEqual 5.72 μs ->2.37 μs
double Min Avg 18.3 μs ->4.0 μs
となっています。
byte型に関してはすべて10倍ほど高速化しています。