@AnnulusGames 2023-12-11
もう一つ、intのSequenceEqualの試行回数を増やすと単体テストを通過しないことがあります。詳しくはまだ調べられていませんが、特定の状況下で通常のLINQと異なる結果を返す可能性があります。