@Akeit0 2023-12-11
すみません、調べてみます。