@Akeit0 2023-12-12
上手く行かなかったのはアライメントの問題でした。Unsafe.ReadUnalignedか同等のilが必要ですね