@AnnulusGames 2023-12-12
Unsafeクラスのdllの追加は他ライブラリとの競合の可能性があるため、Unsafe.ReadUnalignedを使用することは難しいです。UnsafeUtility.ReadArrayElement<T>を用いることで同等の処理が可能でしょうか...?