@Akeit0 2023-12-12
ReadArrayElementはunalignedではないので、IL書くしかないですね。
Unity上で作るなら、Mono.Cecilで書き換える感じです。