@AnnulusGames 2023-12-12
Mono.Cecilの利用は安定性やメンテナンス性の面から可能な限り避けたいと考えています。
現状Min/Maxは問題なく動作するようなので、ひとまずそれら2つをPRに含めるのが良いかと思うのですがいかがでしょうか。