@Akeit0 2023-12-12
Unsafe.ReadUnalignedがなくてもとりあえず大丈夫そうです。