@Akeit0 2023-12-13
問題解消しました。
Average以外ほぼすべてのCore関数をサポートされるVectorの長さで分岐するようにしました。
テストもVector2用のものを追加するなどした上でうまく行っています。