@AnnulusGames 2023-12-13
こちらでも問題なく動作することを確認したのでマージしました。ありがとうございます...！